some docs and refactoring
