+++ distro.js
@@ -3,15 +3,15 @@
     // import WC from './src/index';
     let WC = require('./src/wc'),
 
-    // import on from './extentions/on';
-    // import render from './extentions/render';
-    // import data from './extentions/data';
-    // import template, { templateFragment} from './extentions/template';
+    // import on from './extensions/on';
+    // import render from './extensions/render';
+    // import data from './extensions/data';
+    // import template, { templateFragment} from './extensions/template';
 
-    evts = require('./src/extentions/events'),
-    render = require('./src/extentions/render'),
-    data = require('./src/extentions/data'),
-    templates = require('./src/extentions/template');
+    evts = require('./src/extensions/events/events'),
+    render = require('./src/extensions/render/render'),
+    data = require('./src/extensions/data/data'),
+    templates = require('./src/extensions/template/template');
 
     evts.on(WC);
     evts.trigger(WC);
