+++ src/extensions/events/events.js
@@ -36,16 +36,16 @@ export function on(WC) {
     // Binding to the native Web Component lifecycle methods
     // causing them to trigger relevant callbacks.
     Object.keys(natives).forEach(key => {
-        WC.extentions[natives[key]] = function () {
+        WC.extensions[natives[key]] = function () {
             trigger.call(this, key, ...arguments);
         };
     });
 
-    WC.extentions.on = onEvent;
+    WC.extensions.on = onEvent;
 }
 
 export function trigger(WC) {
     if (WC.missingDeps('trigger', []).length) { return; }
-    WC.extentions.trigger = trigger;
+    WC.extensions.trigger = trigger;
 
 }
