+++ src/extensions/render/domingo.js
