+++ src/extensions/template/template.js
@@ -1,7 +1,7 @@
 export function template(WC) {
 
     if (WC.missingDeps('template', ['on']).length) { return; }
-    WC.extentions.on('created', function () {
+    WC.extensions.on('created', function () {
         var templateEl;
 
         Object.defineProperty(this, 'template', {
@@ -15,7 +15,7 @@ export function templateFragment(WC) {
 
     if (WC.missingDeps('template', ['on']).length) { return; }
 
-    WC.extentions.on('created', function () {
+    WC.extensions.on('created', function () {
         Object.defineProperty(this, 'templateFragment', {
             get: () => document.importNode(this.template.content, true),
             set: (templ) => console.error('templateFragment is not settable')
