+++ src/register.js
@@ -8,7 +8,7 @@ export default (name, config = {}) => {
 
     proto = isString(proto) ? document.createElement(proto) : proto;
 
-    proto = util.protoChain(WC.extentions, proto);
+    proto = util.protoChain(WC.extensions, proto);
 
     return document.registerElement(name, {
         prototype: Object.create(proto, {}),
