+++ src/wc.js
@@ -1,6 +1,6 @@
 import register from './register';
 
-let extentions = {},
+let extensions = {},
 
     WC = Object.create(null, {
         register: {
@@ -10,20 +10,20 @@ let extentions = {},
             value: register.bind(WC)
         },
 
-        // A dictionary of applied extentions.
+        // A dictionary of applied extensions.
         // Every applied extention will be available to each Winston Churchill
         // component in the app.
-        extentions: {
+        extensions: {
             enumerable: true,
-            get: () => extentions,
-            set: (ext) => new Error('Cannot overwrite `extentions` property on WC')
+            get: () => extensions,
+            set: (ext) => new Error('Cannot overwrite `extensions` property on WC')
         },
 
         missingDeps: {
             enumerable: true,
             get: () => function (extention, deps) {
                 let missings = deps.reduce((missing, curr) => {
-                    if (!WC.extentions[curr]) {
+                    if (!WC.extensions[curr]) {
                         missing.push(curr);
                     }
                     return missing;
